Reorganized CLI files.
